+++ src/Psl/Collection/Vector.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace Psl\Collection;
 
-use Psl;
 use Psl\Dict;
 use Psl\Iter;
 use Psl\Vec;
+
+use function array_key_exists;
+use function count;
 
 /**
  * @template T
@@ -104,11 +106,13 @@ final class Vector implements VectorInterface
      * Get the number of items in the current `Vector`.
      *
      * @psalm-mutation-free
+     *
+     * @return int<0, max>
      */
     public function count(): int
     {
-        /** @psalm-suppress ImpureFunctionCall - conditionally pure */
-        return Iter\count($this->elements);
+        /** @var int<0, max> */
+        return count($this->elements);
     }
 
     /**
@@ -142,7 +146,7 @@ final class Vector implements VectorInterface
      *
      * @param int $k
      *
-     * @throws Psl\Exception\InvariantViolationException If $k is out-of-bounds.
+     * @throws Exception\OutOfBoundsException If $k is out-of-bounds.
      *
      * @return T
      *
@@ -150,7 +154,9 @@ final class Vector implements VectorInterface
      */
     public function at(string|int $k): mixed
     {
-        Psl\invariant($this->contains($k), 'Key (%s) is out-of-bounds.', $k);
+        if (!array_key_exists($k, $this->elements)) {
+            throw Exception\OutOfBoundsException::for($k);
+        }
 
         return $this->elements[$k];
     }
@@ -164,8 +170,7 @@ final class Vector implements VectorInterface
      */
     public function contains(int|string $k): bool
     {
-        /** @psalm-suppress ImpureFunctionCall - conditionally pure */
-        return Iter\contains_key($this->elements, $k);
+        return array_key_exists($k, $this->elements);
     }
 
     /**
@@ -383,10 +388,8 @@ final class Vector implements VectorInterface
      *
      * `$n` is 1-based. So the first element is 1, the second 2, etc.
      *
-     * @param $n The last element that will be included in the returned
-     *  `Vector`.
-     *
-     * @throws Psl\Exception\InvariantViolationException If $n is negative.
+     * @param int<0, max> $n The last element that will be included in the returned
+     *                       `Vector`.
      *
      * @return Vector<T> A `Vector` that is a proper subset of the current
      *                   `Vector` up to `n` elements.
@@ -426,10 +429,8 @@ final class Vector implements VectorInterface
      *
      * `$n` is 1-based. So the first element is 1, the second 2, etc.
      *
-     * @param int $n The last element to be skipped; the $n+1 element will be the
-     *               first one in the returned `Vector`.
-     *
-     * @throws Psl\Exception\InvariantViolationException If $n is negative.
+     * @param int<0, max> $n The last element to be skipped; the $n+1 element will be the
+     *                       first one in the returned `Vector`.
      *
      * @return Vector<T> A `Vector` that is a proper subset of the current
      *                   `Vector` containing values after the specified `n`-th element.
@@ -471,11 +472,9 @@ final class Vector implements VectorInterface
      * The returned `Vector` will always be a proper subset of this
      * `Vector`.
      *
-     * @param int $start The starting key of this Vector to begin the returned
-     *                   `Vector`.
-     * @param null|int $length The length of the returned `Vector`
-     *
-     * @throws Psl\Exception\InvariantViolationException If $start or $len are negative.
+     * @param int<0, max> $start The starting key of this Vector to begin the returned
+     *                           `Vector`.
+     * @param null|int<0, max> $length The length of the returned `Vector`
      *
      * @return Vector<T> A `Vector` that is a proper subset of the current
      *                   `Vector` starting at `$start` up to but not including the